+++ lab_09_2/my_getline.h
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <stdbool.h>
 
-int my_getline(char **lineptr, size_t *n, FILE *stream);
+ssize_t my_getline(char **lineptr, size_t *n, FILE *stream);
 bool end_of_line(const char *line, const int sym_count);
 void sym_copy(char *line, const char *symbols, size_t len);
 int prepared_line(char **lineptr, size_t *n, FILE *stream);
